third_party/blink/web_tests/platform/win11-arm64/external/wpt/webxr/hit-test/ar_hittest_source_cancel.https-expected.txt

This is a testharness.js-based test.
[FAIL] Ensures hit test source cancellation works when the session has not ended. - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures hit test source cancellation works when the session has not ended. - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures transient input hit test source cancellation works when the session has not ended. - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures transient input hit test source cancellation works when the session has not ended. - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures hit test source cancellation works when the session has ended - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures hit test source cancellation works when the session has ended - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures transient input hit test source cancellation works when the session has ended - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] Ensures transient input hit test source cancellation works when the session has ended - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
